In the database, the price type was changed from int to decimal

Related work items: #38

+++ MyCar.DataAccess/Models/Refuelling.cs
@@ -9,9 +9,9 @@ public class Refuelling
     public int Odometer { get; set; }
     public int FuelId { get; set; }
     public Fuel Fuel { get; set; }
-    public int Price { get; set; }
-    public int TotalCost { get; set; }
-    public int Quantity { get; set; }
+    public decimal Price { get; set; }
+    public decimal TotalCost { get; set; }
+    public decimal Quantity { get; set; }
     public bool IsFilled { get; set; }
     public string Station { get; set; }
 }
