+++ MyCar.Service/ViewModels/Refuellings/RefuellingGetViewModel.cs
@@ -12,9 +12,9 @@ public class RefuellingGetViewModel
     public int Odometer { get; set; }
     public int FuelId { get; set; }
     public FuelGetViewModel Fuel { get; set; }
-    public int Price { get; set; }
-    public int TotalCost { get; set; }
-    public int Quantity { get; set; }
+    public decimal Price { get; set; }
+    public decimal TotalCost { get; set; }
+    public decimal Quantity { get; set; }
     public bool IsFilled { get; set; }
     public string Station { get; set; }
 }
